// app-client/budgetify-app/src/app/models/categories.ts
export interface ICategory {
  name: string;
  type: string;
}

// [
//   {
//     "name":"Food",
//     "type":"expense"
//   },
//   {
//     "name":"Transportation",
//     "type":"expense"
//   },
//   {
//     "name":"Housing",
//     "type":"expense"
//   },
//   {
//     "name":"Education",
//     "type":"expense"
//   },
//   {
//     "name":"Shopping",
//     "type":"expense"
//   },
//   {
//     "name":"Kids",
//     "type":"expense"
//   },
//   {
//     "name":"Entertainment ",
//     "type":"expense"
//   },
//   {
//     "name":"Health and beauty ",
//     "type":"expense"
//   },
//   {
//     "name":"Pet ",
//     "type":"expense"
//   },
//   {
//     "name":"Internet ",
//     "type":"expense"
//   },
//   {
//     "name":"Mobile",
//     "type":"expense"
//   },
//   {
//     "name":"salary",
//     "type":"income"
//   },
  
//   {
//     "name":"debt repayment",
//     "type":"income"
//   },
  
//   {
//     "name":"gifts",
//     "type":"income"
//   },
  
//   {
//     "name":"rental income",
//     "type":"income"
//   },
  
//   {
//     "name":"premium/bonuses",
//     "type":"income"
//   }
  


//]